+++ mcmc/src/Mcmc/Proposal/Hamiltonian/Hamiltonian.hs
@@ -32,7 +32,7 @@
 -- - The HMC proposal acts on 'Positions', a vector of floating point values.
 --   The manipulated values can represent the complete state, or a subset of the
 --   complete state. Functions converting the state to and from this vector have
---   to be provided; see 'HSpec'.
+--   to be provided; see 'HStructure'.
 --
 -- - Even though the proposal may only act on a subset of the complete state,
 --   the prior, likelihood, and Jacobian functions of the complete state have to
@@ -42,7 +42,7 @@
 --
 -- - The points given above have implications on how the HMC proposal is
 --   handled: Do not use 'liftProposalWith', 'liftProposal', or '(@~)' with the
---   HMC proposal; instead use the conversion functions in 'HSpec'.
+--   HMC proposal; instead use the conversion functions in 'HStructure'.
 --
 -- - The gradient of the log target function is calculated using automatic
 --   differentiation.
@@ -55,13 +55,16 @@
 --   burn in.
 module Mcmc.Proposal.Hamiltonian.Hamiltonian
   ( -- * Hamiltonian Monte Carlo proposal
-    HTuningSpec,
-    hTuningSpec,
+    HParams (..),
+    defaultHParams,
     hamiltonian,
   )
 where
 
+import Control.Monad
+import Control.Monad.ST
 import Data.Bifunctor
+import qualified Data.Vector.Storable as VS
 import qualified Data.Vector.Unboxed as VU
 import Mcmc.Algorithm.MHG
 import Mcmc.Proposal
@@ -72,47 +75,83 @@ import qualified Numeric.LinearAlgebra as L
 import Numeric.Log
 import System.Random.MWC
 
--- | Complete tuning specification of the Hamilton Monte Carlo proposal.
+-- | Parameters of the Hamilton Monte Carlo proposal.
 --
--- Includes tuning parameters and tuning configuration.
-data HTuningSpec = HTuningSpec
-  { hMasses :: Masses,
-    hLeapfrogTrajectoryLength :: LeapfrogTrajectoryLength,
-    hLeapfrogScalingFactor :: LeapfrogScalingFactor,
-    hTuningConf :: HTuningConf
+-- If a parameter is 'Nothing', the value in 'defaultHParams' will be used.
+data HParams = HParams
+  { hMasses :: Maybe Masses,
+    hLeapfrogTrajectoryLength :: Maybe LeapfrogTrajectoryLength,
+    hLeapfrogScalingFactor :: Maybe LeapfrogScalingFactor
   }
   deriving (Show)
 
--- | See 'HTuningSpec'.
+-- | Default parameters.
 --
--- Return 'Left' if an error is found.
-hTuningSpec ::
-  Masses ->
-  LeapfrogTrajectoryLength ->
-  LeapfrogScalingFactor ->
-  HTuningConf ->
-  Either String HTuningSpec
-hTuningSpec masses l eps c
-  | any (<= 0) diagonalMasses = eWith "Some diagonal entries of the mass matrix are zero or negative."
-  | nrows /= ncols = eWith "Mass matrix is not square."
-  | l < 1 = eWith "Leapfrog trajectory length is zero or negative."
-  | eps <= 0 = eWith "Leapfrog scaling factor is zero or negative."
-  | otherwise = Right $ HTuningSpec masses l eps c
-  where
-    eWith m = Left $ "hTuningSpec: " <> m
-    ms = L.unSym masses
-    diagonalMasses = L.toList $ L.takeDiag ms
-    nrows = L.rows ms
-    ncols = L.cols ms
+-- - Mass matrix is identity matrix.
+--
+-- - Leapfrog trajectory length is 10.
+--
+-- - A resonable leapfrog scaling factor will be estimated.
+defaultHParams :: HParams
+defaultHParams = HParams Nothing Nothing Nothing
 
-hTuningParametersToTuningSpec ::
-  HTuningSpec ->
+-- Internal. We need two data types here. 'HParams' is exposed to the
+-- user. This one is exposed to the algorithm, and has all values set.
+data HParamsSet = HParamsSet
+  { hMasses' :: Masses,
+    hLeapfrogTrajectoryLength' :: LeapfrogTrajectoryLength,
+    hLeapfrogScalingFactor' :: LeapfrogScalingFactor,
+    hData :: HData
+  }
+  deriving (Show)
+
+-- Check 'HParams'; set parameters; compute 'HData'.
+--
+-- NOTE: This function may sample momenta to find a reasonable leapfrog scaling
+-- factor. Hence, it should be run in the IO monad. However, I opt for a pure
+-- setting here and use a fixed seed.
+fromHParams :: Target -> Positions -> HParams -> Either String HParamsSet
+fromHParams htarget p (HParams mMs mL mE) = do
+  d <- case VS.length p of
+    0 -> eWith "Empty position vector."
+    d -> Right d
+  let defMs = L.trustSym $ L.ident d
+  ms <- case mMs of
+    Nothing -> Right defMs
+    Just ms -> do
+      let ms' = L.unSym ms
+          diagonalMs = L.toList $ L.takeDiag ms'
+      when (any (<= 0) diagonalMs) $ eWith "Some diagonal masses are zero or negative."
+      let nrows = L.rows ms'
+          ncols = L.cols ms'
+      when (nrows /= ncols) $ eWith "Mass matrix is not square."
+      Right ms
+  let hdata = getHData ms
+  l <- case mL of
+    Nothing -> Right 10
+    Just l
+      | l <= 0 -> eWith "Leapfrog trajectory length is zero or negative."
+      | otherwise -> Right l
+  e <- case mE of
+    Nothing -> Right $ runST $ do
+      g <- create
+      findReasonableEpsilon htarget ms hdata p g
+    Just e
+      | e <= 0 -> eWith "Leapfrog scaling factor is zero or negative."
+      | otherwise -> Right e
+  pure $ HParamsSet ms l e hdata
+  where
+    eWith m = Left $ "fromHParams: " <> m
+
+hTuningParametersToHParamsSet ::
+  HParamsSet ->
+  HTuningConf ->
   TuningParameter ->
   AuxiliaryTuningParameters ->
-  Either String HTuningSpec
-hTuningParametersToTuningSpec (HTuningSpec ms l e c) t ts
-  | not nTsOK = Left "hTuningParametersToSettings: Auxiliary variables dimension mismatch."
-  | otherwise = Right $ HTuningSpec msTuned lTuned eTuned c
+  Either String HParamsSet
+hTuningParametersToHParamsSet (HParamsSet ms l e hdata) c t ts
+  | not nTsOK = Left "hTuningParametersToHParamsSet: Auxiliary variables dimension mismatch."
+  | otherwise = Right $ HParamsSet msTuned lTuned eTuned hdataTuned
   where
     d = L.rows $ L.unSym ms
     (HTuningConf tlf tms) = c
@@ -121,9 +160,9 @@ hTuningParametersToTuningSpec (HTuningSpec ms l e c) t ts
        in case tms of
             HNoTuneMasses -> nTs == 0
             _ -> nTs == d * d
-    msTuned = case tms of
-      HNoTuneMasses -> ms
-      _ -> tuningParametersToMasses d ts
+    (msTuned, hdataTuned) = case tms of
+      HNoTuneMasses -> (ms, hdata)
+      _ -> let ms' = tuningParametersToMasses d ts in (ms', getHData ms')
     -- The larger epsilon, the larger the proposal step size and the lower the
     -- expected acceptance ratio.
     --
@@ -136,38 +175,38 @@ hTuningParametersToTuningSpec (HTuningSpec ms l e c) t ts
 
 hamiltonianPFunctionWithTuningParameters ::
   Traversable s =>
-  HTuningSpec ->
-  HSpec s ->
+  HParamsSet ->
+  HTuningConf ->
+  HStructure s ->
   (s Double -> Target) ->
   TuningParameter ->
   AuxiliaryTuningParameters ->
   Either String (PFunction (s Double))
-hamiltonianPFunctionWithTuningParameters tspec hspec targetWith t ts = do
-  tspec' <- hTuningParametersToTuningSpec tspec t ts
-  pure $ hamiltonianPFunction tspec' hspec targetWith
+hamiltonianPFunctionWithTuningParameters tspec htconf hstruct targetWith t ts = do
+  tspec' <- hTuningParametersToHParamsSet tspec htconf t ts
+  pure $ hamiltonianPFunction tspec' hstruct targetWith
 
 -- The inverted covariance matrix and the log determinant of the covariance
 -- matrix are calculated by 'hamiltonianPFunction'.
 hamiltonianPFunctionWithMemoizedCovariance ::
   Traversable s =>
-  HTuningSpec ->
-  HSpec s ->
-  HData ->
+  HParamsSet ->
+  HStructure s ->
   (s Double -> Target) ->
   PFunction (s Double)
-hamiltonianPFunctionWithMemoizedCovariance tspec hspec dt targetWith x g = do
-  phi <- generateMomenta mu masses g
+hamiltonianPFunctionWithMemoizedCovariance tspec hstruct targetWith x g = do
+  phi <- generateMomenta mu ms g
   lRan <- uniformR (lL, lR) g
   eRan <- uniformR (eL, eR) g
-  case leapfrog (targetWith x) massesInv lRan eRan theta phi of
+  case leapfrog (targetWith x) msInv lRan eRan theta phi of
     -- TODO @Dominik (high, feature): Acceptance counts.
     Nothing -> pure (ForceReject, Nothing)
     -- Check if next state is accepted here, because the Jacobian is included in
     -- the target function. If not: pure (x, 0.0, 1.0).
     Just (theta', phi', prTheta, prTheta') -> do
       let -- Prior of momenta.
-          prPhi = exponentialKineticEnergy massesInv phi
-          prPhi' = exponentialKineticEnergy massesInv phi'
+          prPhi = exponentialKineticEnergy msInv phi
+          prPhi' = exponentialKineticEnergy msInv phi'
           r = prTheta' * prPhi' / (prTheta * prPhi)
       accept <- mhgAccept r g
       -- NOTE: For example, Neal page 12: In order for the Hamiltonian proposal
@@ -179,22 +218,22 @@ hamiltonianPFunctionWithMemoizedCovariance tspec hspec dt targetWith x g = do
       -- TODO @Dominik (high, feature): Acceptance counts.
       pure (pr, Nothing)
   where
-    (HTuningSpec masses l e _) = tspec
-    (HSpec _ toVec fromVec) = hspec
+    (HParamsSet ms l e hdata) = tspec
+    (HData mu msInv) = hdata
+    (HStructure _ toVec fromVec) = hstruct
     theta = toVec x
     lL = maximum [1 :: Int, floor $ (0.8 :: Double) * fromIntegral l]
     lR = maximum [lL, ceiling $ (1.2 :: Double) * fromIntegral l]
     eL = 0.8 * e
     eR = 1.2 * e
-    (HData mu massesInv) = dt
 
 hamiltonianPFunction ::
   Traversable s =>
-  HTuningSpec ->
-  HSpec s ->
+  HParamsSet ->
+  HStructure s ->
   (s Double -> Target) ->
   PFunction (s Double)
-hamiltonianPFunction tspec hspec = hamiltonianPFunctionWithMemoizedCovariance tspec hspec (getHData $ hMasses tspec)
+hamiltonianPFunction hparams hstruct = hamiltonianPFunctionWithMemoizedCovariance hparams hstruct
 
 hGetTuningFunction :: Int -> (a -> Positions) -> HTuningConf -> Maybe (TuningFunction a)
 hGetTuningFunction n toVec (HTuningConf l m) = case (l, m) of
@@ -211,37 +250,39 @@ hGetTuningFunction n toVec (HTuningConf l m) = case (l, m) of
 -- | Hamiltonian Monte Carlo proposal.
 hamiltonian ::
   (Eq (s Double), Traversable s) =>
-  HTuningSpec ->
-  HSpec s ->
+  HParams ->
+  HTuningConf ->
+  HStructure s ->
   HTarget s ->
   PName ->
   PWeight ->
   Proposal (s Double)
-hamiltonian tspec hspec htarget n w = case checkHSpecWith (hMasses tspec) hspec of
-  Just err -> error err
-  Nothing ->
-    let -- Misc.
-        desc = PDescription "Hamiltonian Monte Carlo (HMC)"
-        (HSpec sample toVec fromVec) = hspec
-        dim = L.size $ toVec sample
-        pDim = PSpecial dim 0.65
-        -- Vectorize and derive the target function.
-        (HTarget mPrF lhF mJcF) = htarget
-        tF y = case (mPrF, mJcF) of
-          (Nothing, Nothing) -> lhF y
-          (Just prF, Nothing) -> prF y * lhF y
-          (Nothing, Just jcF) -> lhF y * jcF y
-          (Just prF, Just jcF) -> prF y * lhF y * jcF y
-        tFnG = grad' (ln . tF)
-        targetWith x = bimap Exp toVec . tFnG . fromVec x
-        ps = hamiltonianPFunction tspec hspec targetWith
-        hamiltonianWith = Proposal n desc PSlow pDim w ps
-        -- Tuning.
-        tconf@(HTuningConf _ tms) = hTuningConf tspec
-        ts = case tms of
-          HNoTuneMasses -> VU.empty
-          _ -> massesToTuningParameters $ hMasses tspec
-        tuner = do
-          tfun <- hGetTuningFunction dim toVec tconf
-          pure $ Tuner 1.0 ts tfun (hamiltonianPFunctionWithTuningParameters tspec hspec targetWith)
-     in hamiltonianWith tuner
+hamiltonian hparams htconf hstruct htarget n w =
+  let -- Misc.
+      desc = PDescription "Hamiltonian Monte Carlo (HMC)"
+      (HStructure sample toVec fromVec) = hstruct
+      dim = L.size $ toVec sample
+      pDim = PSpecial dim 0.65
+      -- Vectorize and derive the target function.
+      (HTarget mPrF lhF mJcF) = htarget
+      tF y = case (mPrF, mJcF) of
+        (Nothing, Nothing) -> lhF y
+        (Just prF, Nothing) -> prF y * lhF y
+        (Nothing, Just jcF) -> lhF y * jcF y
+        (Just prF, Just jcF) -> prF y * lhF y * jcF y
+      tFnG = grad' (ln . tF)
+      targetWith x = bimap Exp toVec . tFnG . fromVec x
+      hparams' = either error id $ fromHParams (targetWith sample) (toVec sample) hparams
+      ps = hamiltonianPFunction hparams' hstruct targetWith
+      hamiltonianWith = Proposal n desc PSlow pDim w ps
+      -- Tuning.
+      (HTuningConf _ tms) = htconf
+      ts = case tms of
+        HNoTuneMasses -> VU.empty
+        _ -> massesToTuningParameters $ hMasses' hparams'
+      tuner = do
+        tfun <- hGetTuningFunction dim toVec htconf
+        pure $ Tuner 1.0 ts tfun (hamiltonianPFunctionWithTuningParameters hparams' htconf hstruct targetWith)
+   in case checkHStructureWith (hMasses' hparams') hstruct of
+        Just err -> error err
+        Nothing -> hamiltonianWith tuner